+++ gazebo/math/SignalStats.cc
@@ -142,11 +142,6 @@ SignalStats::SignalStats(const ignition::math::SignalStats &_s)
        iter != _s.Map().end(); ++iter)
   {
     this->InsertStatistic(iter->first);
-  }
-
-  for (std::map<std::string, double>::const_iterator iter = _s.Map().begin();
-       iter != _s.Map().end(); ++iter)
-  {
     this->InsertData(iter->second);
   }
 }
@@ -271,10 +266,10 @@ ignition::math::SignalStats SignalStats::Ign() const
   ignition::math::SignalStats result;
 
   for (auto const &statistic : this->dataPtr->stats)
+  {
     result.InsertStatistic(statistic->ShortName());
-
-  for (auto const &statistic : this->dataPtr->stats)
     result.InsertData(statistic->Value());
+  }
 
   return result;
 }