+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Gazebo 8.x.x (2017-xx-xx)
 
+1. Add Gazebo math helper functions to convert to and from Ignition Math
+   objects.
+    * [Pull request #xxxx](https://bitbucket.org/osrf/gazebo/pull-request/xxxx)
+
 1. Add video recording of user camera. This change added an optional
    dependency on libavdevice>=56.4.100 for linux systems. When installed,
    libavdevice will allow a user to stream a simulated camera to a video4linux2